requirements.txt
python-dotenv>=1.0.0
mcp>=1.8.1
pyautogui>=0.9.54
opencv-python>=4.10.0.24
onvif-zeep>=0.2.12
getmac>=0.8.0
python-dateutil>=2.8.2
requests>=2.25.1
Pillow>=8.3.0
numpy>=1.19.0
